F1632 eaf with mig29 rwr (#18)
Made a clone of the F16 B32 EAF called "F16 B32EAF" (Without the space between 32 and EAF) which will show up as a mig 29 on the RWR, used for UOAF campaign option. Did NOT replace the squadrons in the stock theater. Idea would be to save camp 1, then mass replace all the Egyptian viper squadrons with this new clone